api_repo/requirements.txt
fastapi
uvicorn
nbconvert
papermill
jupyter
gitpython

numpy
scipy
pandas
matplotlib
laspy
rasterio
shapely